Found a counter-example of push_iter.v
In the algorithm, when we check if the new result ds1 of a node equals to
its old one ds2, we only check if their heads are the same (recall we
represent dominators by a list). To show that equivalent heads imply
that two lists of dominators equal, the invariant that must hold is
  given the current dominators doms, at each iteration,
  if dom[x] = Some dms_x, then 
  forall dom[y] = Some _++x::tail, tail = dms_x

Unfortunately, this is not true for push. Here is a simplied
counter-example (from idfsquad of "Finding Dominators in Practice"):
     
define void @foo(i1 %b) {
%n8: br i1 %b, label %n7, label %n5

%n7: br i1 %b, label %n4, label %n6
%n4: br i1 %b, label %n1, label %n3

%n6: br i1 %b, label %n5, label %n2
%n3: br label %n2

xn5: br label %n6
%n2: br label %n3

%n1:
  ret void
}

Nodes are rpo-numbered.

At some iteration,
doms[2] = [6;7;8]
doms[6] = [8]

2 is the successor of 6.
old = doms[2]
new = old /\ (doms[6] \/ {6}) = [6;8]

head [6;8] = head [6;7;8], but [6;8] <> [6;7;8] !!

Thie leads to wrong dominator results.

worstcases/*.ll are the 4 examples from "Finding Dominators in
Practice". Pull+RPO is correct for all of them. Push+RPO is wrong for
idsquad.ll, but correct for others. 
  sh run.sh <name>  # compile, gen cfg-dot, dom-analysis, diff



git-svn-id: https://webdav.seas.upenn.edu/svn/sol/vol@4000 eef5ea85-cc2e-0410-bd84-0fae1f012da8