* Added counter-examples for both our pull-based algorithm and the original pull-based algorithm (Figure3 of "A Simple, Fast Dominance Algorithm"). The original algorithm is a round-robin iteration. Our pull is a worklist iteration.
counter-example1.ll only fails push. counter-example2/3 fail both of
them. The DFSs of counter-example2/3 can be taken as mirrors of the DFS
of counter-example1.

This concludes that we cannot simply check equality of two lists of
dominators by checking their heads. Note that the orinigal algorithm
uses a map to store candidate idom, checking if cadidate idoms equal is
the same to checking if heads of lists used in our algorithm equal.

Intuitively, if we could prove that (see r4000)
  given the current dominators doms, at each iteration,
    if dom[x] = Some dms_x, then
  forall dom[y] = Some _++x::tail, tail = dms_x,
we would prove that the algorithm only needs one pass to stop, because
it is equivalent to saying that once a set of dominators of a node is
computed, they are not changed. We know experiments show that the
algorithm is not linear. So, it cannot be provable.

The counter-examples are not easy to find. olden/spec95/2k/2k6 do not
contain such programs. The complex worst cases of "Finding dominators in
practice" do not trigger the error either, because the DFS those cases
generate wont fail pull, but only push. 

* This does not mean we cannot fix the bug.
  1) "intersect" should not stop when we find the first equivalent
     nodes, but should continue at the end of two lists. This is exactly
     finding the same elements from two sorted lists.
  2) The algorithm only checks if old = new after intersection, actually
     the modified intersection has already known the answer while
     traversing the two sorted lists. So, we can simply reuse the answer
     w/o re-checking equivalence of the two lists.

  As for complexity, the fixed version should be in the same complex
  class as the buggy one. intersection is O(D), here D is the average
  depth of DFS, log(N), no matter we stop at the first equal elements or
  go through all of them. But, in practice, they should be different...

* The bug we found only happen on irreducible CFGs. However, checking if
  a CFG is reduciable could be as hard as checing SCC, which is as hard
  as computing dom-trees.



git-svn-id: https://webdav.seas.upenn.edu/svn/sol/vol@4002 eef5ea85-cc2e-0410-bd84-0fae1f012da8